+++ ex03/src/main/java/org/zerock/domain/BoardVO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class BoardVO {
     
-    private Long bno;
+    private int bno;
     private String title, content, writer;
     private Date regdate, updateDate;
